move history workflow extraction from API controller to service layer
